Fixes Issue 39 - Crash at zfs_vnop_reclaim
When we reclaim a node, it's possible that there isn't any ZFS data
associated with it (either because it wasn't there in the first place,
or because it's been subsequently collected). These were causing panics.
If we detect whether the node has been used or not, and select
appropriately, then we don't panic.

See: http://code.google.com/p/maczfs/issues/detail?id=39

Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>