Added testcases
These are helper scripts, assuming that private keys are set up to
automatically log into a machine called (or aliased) to 'target',
and that the user has a suitable sudo environment in order to execute
passwordless sudo commands.

Alternatively, they may be copied and pasted onto the local machine,
via a target kicked off with the zfs-target-installer script found in
the support directory.

This assumes the files are loaded into /target, and that the PATH and
DYLD_LIBRARY_PATH are set up to be able to execute commands from the
/target path.

See: http://code.google.com/p/maczfs/wiki/KernelDebugging

Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>