Updated build environemnt variables for 10.5/10.6
The build environment used a ZFS_LEOPARD_ONLY flag which
was being used for 10.5 builds. Remove this from the kernel,
and have the flag auto-detected for 10.5 builds based on the
compiler version.

This was implemented incompletely before; the auto-detection
needed to be placed higher up (since a number o fheader files
in zfs_context also used this) and the zfs.kext included a
duplicate definition of this environment variable which prevented
use on 10.6 systems.

Tested on Intel 10.5 and Intel 10.6