Added build configuration for LLVM
This adds an LLVM build configuration (using the Xcode 3.2 LLVM 1.5)
as well as refactoring the build to take advantage of LLVM.
Since LLVM is not on 10.5, we can't make this the default for all
projects, though it is about 25-30% faster at building.

In addition, it gives good static analysis advice, which makes it
easier to scan code for problems.