Provide a way of disabling 'stalk' messages
The zfs_znode.c has a znode_stalker function, which
generates voluminous input when compiled with ZFS_DEBUG
(which the debug builds are enabled with). In addition,
this will build up a queue of events in the 'z_stalker'
list, which could grow large. Provide a sysctl variable
vfs.maczfs.debug.stalk to turn this off, and have it
disabled by default..

Moved Mac-specific source to /usr/src/maczfs, and kernel-
specific code into /usr/src/maczfs/kernel, in an attempt
to clean up some of the code.