Allow vnode_t to be used instead of specific ifdefs
The OpenSolaris codebase uses vnode_t as struct vnode, whereas
Mac OS X uses vnode_t as struct vnode *. So the types aren't
compatible, and we can't use a typedef to hook them up.

Instead, inspired by the 10a286 change, we use #define to replace
all occurances of vnode_t with struct vnode, such that in the
code, we refer to the right piece. However, we need to be careful
about the places where the standard system libraries are brought in;
to do that, hooks for the maczfs_{file,ubc,mount,libc) have been
provided which undo that typedef on the fly.