Add gzip support for filesystem compression
Using the built-in Kernel libraries, add gzip compression
to the ZFS support. This was made possible thanks to the
10a286 bits that Apple released under the terms of the CDDL.

The default compression is still off, and on will default to
lzjb. gzip is an alias for gzip-6; you can also select gzip-1..9
as compression levels. Using gzip compression will give better
compression; however, the lzjb is faster for most workloads.
Compression of 1.5 times for source code with lzjb is average;
gzip-1 will give 2.5 times and gzip-9 will give 3 times. However,
the additional procesing speed to compress means that gzip-9 is
only really suitable for archiving purposes. The read decompression
is similar across all compression levels.

Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>