Increase priority of the zfs.fs filing system
The original Apple zfs.fs filing system has a priorty
of 1000 (the read-only one may have had a number greater
than that).

However, when ZFS was dropped from 10.6, Apple planted a
nofs.fs which took priority over the (Apple-shipped) ZFS
implementation. The nofs.fs has a priorty of 900.

Leapfrog the nofs.fs back again so that we get priority,
but without having to do any changing of the local file
system to make it happen.